+++ packages/base/src/panelview/rightpanel.tsx
@@ -6,7 +6,7 @@ import {
 import { SidePanel } from '@jupyterlab/ui-components';
 
 import { IControlPanelModel } from '../types';
-import { FilterPanel } from './components/Filter';
+import { FilterPanel } from './components/filter-panel/Filter';
 import { ControlPanelHeader } from './header';
 import { ObjectProperties } from './objectproperties';
