+++ packages/base/style/filterPanel.css
@@ -1,0 +1,43 @@
+.jp-gis-filter-main {
+  display: flex;
+  flex-direction: column;
+  padding: 10px;
+  gap: 1rem;
+}
+
+.jp-gis-filter-list {
+  display: flex;
+  flex-direction: column;
+}
+
+.jp-gis-filter-button-container {
+  display: flex;
+  justify-content: flex-end;
+}
+
+.jp-gis-filter-main > .jp-Dialog-button {
+  width: fit-content;
+  align-self: flex-end;
+}
+
+.jp-gis-filter-select-container {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.jp-gis-filter-row {
+  display: flex;
+  justify-content: space-around;
+  width: 100%;
+  gap: 0.25rem;
+}
+
+.jp-gis-filter-row > select {
+  text-transform: capitalize;
+}
+
+.jp-gis-filter-row > option {
+  text-transform: capitalize;
+  font-size: var(--jp-ui-font-size1);
+}
